Added a command for monthly group sales reset. Changed points data type from int to double (database)

+++ database/migrations/2019_08_13_172455_create_genealogy_match_points_table.php
@@ -18,10 +18,10 @@ class CreateGenealogyMatchPointsTable extends Migration
             $table->unsignedBigInteger('genealogy_id');
             $table->bigInteger('matches')->default(0);
             $table->bigInteger('flushed_out_matches')->default(0);
-            $table->bigInteger('product_points')->default(0);
+            $table->double('product_points')->default(0);
             $table->bigInteger('incentives_points')->default(0);
-            $table->bigInteger('left_group_sales_points')->default(0);
-            $table->bigInteger('right_group_sales_points')->default(0);
+            $table->double('left_group_sales_points')->default(0);
+            $table->double('right_group_sales_points')->default(0);
             $table->timestamps();
 
             $table->foreign('genealogy_id')->references('id')->on('genealogies')->onDelete('cascade');
