+++ database/migrations/2019_08_13_172748_create_user_product_logs_table.php
@@ -17,7 +17,7 @@ class CreateUserProductLogsTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('product_id');
-            $table->bigInteger('points')->default(0);
+            $table->double('points')->default(0);
             $table->bigInteger('total')->default(0);
             $table->string('remarks')->default('None');
             $table->timestamps();
